I now have all the basic function for matrix3 in place. They all need optimising though.
I have begun the rename of all the functions to bring a consistancy
to the naming. This will be followed by a rename of the packages.
Apparently we need to be careful with this so git doesnt lose the
history.

Also going to import some functions from cepl-vector3 into
cepl-matrix3 so the full path is not required, this will make
the functions tidier. I'll will then have to go back and realign
these as they are currently a bugger to read.