A tidy up has been had!
base has been tidied up a bit and a new function added though it
may be wrong, I have a todo to go find out.

Matrix functions have been laid out so it is easier to read the
code and infer what is happening. Turns out I confused myself
with matix indexing as I was thinking x&y and row&columns.

I have started playing properly with macros. I wrote the math-macro
package with the idea of abstractiong away the repetitive parts of
the vector and matrix functions.
The first macro in this package is to help with the vector functions
as they are nearly identical across dimentions

It can expand the following:
;; (apply-across-elements make-vector3 ((vc-a vector-a)) 3
;;   (* vc-a b)))

Into this:
;; (MAKE-VECTOR3 (* (AREF VECTOR-A 0) B)
;;               (* (AREF VECTOR-A 1) B)
;; 	         (* (AREF VECTOR-A 2) B))

Thats all for now
Ciao!