Righto, this is the first commit where I've got some code to run. At the moment I am trying out the cl-opengl library in order to see what kinds of features I'm going to need to add to cepl to make it pleasent to code with.
This is my first experience writing anything with opengl so
its been fun trying to get a handle on the lisp code while at
the same time learning opengl.

I'm using the fantastic tutorials at www.arcsynthesus.org/glut
and this commit has the first hash at the first tutorial.

The code is butt-ugly as it is aimed at being as close to the c++
as possible without using the nicer functions in cl-opengl.
The next commit will tidy this up.