playing with ideas for the first opengl abstractions
I'm writing down whats in my noggin and then having a go at
writing the pseudo functions.

I will have cepl-gl as the gl library, the idea being to only
expose the parts of opengl wort hkeeping while not being enough
of a dictator to obstruct anyone using them.

All cepl's features though will be based on modern opengl.
This will be very important if I ventually get to the point
of handling openglES, but that is a long LONG way off yet.