Right after the last commit I broke some of the tutorial demos. I have now made a new pakcage called defunct which holds the old versions of some functions. This will keep these demso running until I rewrite them at which point the origianls and the defuct package will be consigned to history.
I have also modified make-shader so that it has a restart case
on the event of a failed compilation. This allows the use to
tweak and compile in shader code without fear of killing the
program.

Oh and I have renamed 'restartable' to 'continuable' and it
implemented a continue, not a restart. Don't know why the orginal
author named it like that (though I am very glad they did as it
allowed be to get my head around conditions!)

Right, the next job is to focus on the opengl 'program' and how
I will wrap that up.
My current thinking is to have make-program return and lambda
which takes the uniforms as arguments and a stream which will be
used as the 'attributes' to the shaders (bah mixed up naming
conventions!)

Peace