WARNING WILL NOT COMPILE Ok this commit really is just so I can then go make serious changes to cepl:gl and have some record of what I have been fiddling with.
I have bought a copy of 'let over lambda' recently and the
effects are starting to be obvious...more macros!

Not always a good thing, but the experience is fun so they
are in for now. Also I am starting to be able to make some
opinionated choices about how I want to interface with OpenGL.

base-macros is a home for macros I am using across the board

base-maths is base.lisp renamed

cepl-gl is not in its full form yet so this version will not compile
(in fact this will be true for much of this commit!)

cepl-time is home for the time functions I have started making.
Thesse a really to try and abstract areas of the game loop so we
dont have to have some kind of special loop macro just for games.
That always rubs me up the wrong way.

cepl-utils is expanding with handy, if slow, functions

cepl.asd is not correct yet

cepl-package is modified

matrix3 and 4 have been tweaked as packages have been renamed