*WARNING: NEW STUFF ALMOST CERTAINLY WILL NOT WORK*
Alright! This is a very exciting one as I have been powering
away at the cepl-gl package and i think I may have all the
parts finally written. They have not been run yet so will
definitely fail BUT all the main abstractions around cl-opengl
are there.

My next step is to write some demos to make sure all the features
work (and fix them when they invariably don't!).

Ok so the main ideas in cepl-gl (cgl from now on)

bind-buffer, use-program and bind-vertex-array are all memoized
functions which only evaluate if the value you pass in is
different from the value passed in on the last call.
This minimises the number of calls and thus allows it to be
used more liberaly without concern for the cost of the bind.
Of course this adds a little overhead but the idea is that the
user will never really have to use these functions themselves.
A few functions have been shadowed to use this functionality
but it's still early days here.

I have extended the functionality of the gl-arrays provided by
cl-opengl. They were good but provided functionality for all
versions of opengl. By forgoing this and focusing on opengl 2.3+
I was able to use the binding functions to autogenerate the
attrib-pointer calls used to make the vaos.

Oh I have also dropped the idea of streams=vao
renderable=vao+details in favour of a naming convention closer
to the opengl spec. There are soem pretty cool changes though:
* you define data structures using define-gl-array-format and
  use it to create gl-arrays. (you can use the
  destructured-populate to make filling arrays MUCH tidier)
* You use multi-populate-buffer to fill a buffer with a list of
  arrays. Arrays using the gl-array-formats are interleaved and
  the arrays a packed in one after another in the buffer.
  multi-populate-buffer spits out a buffer-format list which is
  can be used to autogenerate vaos
* make-vao takes a list telling it what buffer (and where in the
  buffer) to take the data for each attribute from. By using
  multi-populate-buffer you are given this automatically.
  It will then create a vao and spit out the vao-id
* streams are structs which contain details for drawing, namely:
  vao-id, start, length, element-type (nil unless making using
  an element-array-buffer) and a draw style (defaults to
  :triangles).
* Programs are a big leap from the standard.
  when you call make-program you pass in a list of shaders
  make-program links them and produces a program-id. This is then
  used to extract the uniforms from the shader program. The
  uniform names are then lispified ("aVal_b_c"->":AVAL-B-C")
  and used as a hash key for the function used to update that
  particular type of uniform.
  make-program then makes a closure which takes a list of streams
  and an optional &key list of uniforms to update. Once it has
  updated any uniforms it has had passed in it then loops through
  the streams and calls draw-arrays (or draw-elements) on them.

Programs seem crazy but they are essentially treating glsl
programs as a function. You give the function streams and it
draws them. The function remembers the uniforms it was provided
previously so you only need to pass uniforms details if you need
to update them.

There is LOOOADS of room for optimisation in cepl-gl, it is just
a first hash at the abstractions I think will make writing things
in opengl fast and fun. The opengl state is a bitch to keep track
of in your head and so I try to limit that as much as possible
by making the calls look like concepts that are familiar in lisp
(closures, lists, structs and arefs).

Other changes are pretty trivial. Ok so now I am off to try and
use this stuff...wish me luck!