*ITS STILL NOT WORKING!* I've made some progress here that I'm very happy with but I've also run into a limitation of the macro system defined in cl-opengl and really need to scrap some of it and rework it.
It will mean doing some things at run time, but they are the kinds
of operations that are run very infrequently so trying to optimise
the situation through the macros is only increasing complexity
without providing any messured benefit.
I'd like to be very clear though, I like macros cl-opengl put in
because for some of the situations in fixed function rendering it
would have been a really boon, but alas not for modern opengl work.

Most of the commit is bug fixing and tidying up but as I'm about
to go on a code-culling-rampage I thought I better commit.