IT LIVES! This is the working push of the new opengl abstractions. I've been working through the old arc-synthesis tutorials again both to ensure cepl-gl is bug free and also to make sure it is pleasent to use. In both cases it is certainly getting closer!
cepl-gl
-------
I have done some serious reworking here to get things to work and
feel right. The latter is something that still needs work and the
state of the package is still not great in that alot of things
aren't exported yet.

The flow of getting data into opengl buffers is now much simpler
than before.
First off you can define 'attribute formats' these are the
data-types that our vertex stream will be made off.
so:
(define-attribute-format just-position :type :float
                                       :components (x y z w))
gives you a datatype called 'just-position' which will be passed
to your shader/s as a float-4 vector.

Seen as we have to get data into gl-arrays first so we can send
them to buffers, we can use the attribute format we just made as
an array type so: (alloc-gl-array 'just-position 10) will make a
gl-array (note: gl-arrays are just pretty wrappers around c-arrays)
of 10 vectors each with an x,y,z & w coordinate.

Now often we want to interleave our data as it makes it faster for
the gpu to read...unfortunately working out the strides and offsets
is a royal pain in the bum, so I've made it really easy with
attribute formats.
(define-interleaved-attribute-format position-and-color
            		 (:type :float :components (x y z w))
			 (:type :unsigned-char :components (r g b)))
The code above will make an attribute format with interleaved data.
we can make arrays with it just like before

If you need to quickly squirt data into these arrays you may find
writing loops and using aref-gl a bit repetative so you can use
'destructuring-populat
This takes an array and a list of data and populates the array
using the data in the list.

 (setf *pos-and-color-array* (alloc-gl-array 'position-and-color 3))
 (destructuring-populate *pos-and-color-array*
                '((( 0.0   0.2  0.0  1.0) (255   0   0))
      		  ((-0.2  -0.2  0.0  1.0) (0   255   0))
		  (( 0.2  -0.2  0.0  1.0) (0     0 255))))

Notice how the data in the list mirrors the structure of the
attribute format from above.

Ok so now you have data in an array and you need it in a buffer.
All you do is call buffer-data passing the array and buffer-id.
cepl's buffer-data is different from the standard opengl one as it
returns a list containing the layout  of all the arrays in the
buffer with all the stride information, offsets and such calculated.
The reason for this is that it can be passed straight into the
'make-vao' functions to create a vao. You can mix and match data
from different buffers if you like but it will certainly be faster
if the data is in one buffer wherever possible.

vaos can be used in making streams. Streams are something that are
used constantly in opengl but don't have their own datatype. In
cepl a stream is a vao with a start and end index which are used
when rendering.

another area that is substantially different from standard opengl
in syntax, is the opengl programs.
Programs are compiled shaders which are used to encapsulate the
pipeline.
I felt that rather than this being just a number to be passed to
'use-program' it would be much more logical to treat this as a
function which ran on the gpu. In lisp, we deal a lot with lambdas
so when you run cepl's 'make-program' and pass in a list of shaders,
you get given back a lambda.
When you call this function, passing in a list of stream, it will
render the data in those streams using the gpu pipeline.
As well as streams of data, shader programs also use uniforms.
These can also easily be passed in using the uniform name as a key
so to pass in the 'camera-matrix' uniform you can do the following
(funcall *program-lambda* (stream1 stream2)
	                  :camera-matrix (gen-camera-matrix))
The uniforms are stored in the lambda's lexical scope between calls
so you only need to pass in a uniform when you need to update it..
NOT every loop.

---------------------

Also in cepl-gl I fixed a STUPID bug in the matrix uniform code.

---------------------

cepl-time.lisp was missing an 'in-package' so that is added now.

---------------------

package.lisp has obviously been updated but cepl-gl is still messy
and incomplete...I will be looking into this soon.
The user should Never have to use cl-opengl

---------------------

Some files have been changed and added in arc-tuts but this whole
directory is pretty much redundant now and will be removed in the
next commit.

---------------------

the 'examples' directory now contains the SDL versions of the
arc-synthesis tutorials I have done thus far. It is where I am
playing with the features and seeing how they look in 'real' code.

The naming of the files is crap and will be fixed in the next commit

---------------------

the 'experiments' directory is more a stash of my mental droppings
rather than real experiments, I'm commiting these files for the
record but they may be deleted in the next commit.

---------------------

Thats all folks, this took WAAYY to long but I've learnt alot so
thats a relief at least!
Hope your all well
Ciao