Added a new demo showing world to camera transforms. very happy with this and am now in a position to having multiple 'objects' in a scene and rotate, scale and position them as I choose.
It's all still simple stuff but it allows me to start seeing where
I need to make the experience for the users as clean as possible and
where I really need to work.

It seems like a good time to go and review some of the functional
programming styles and read some articles on how they could pertain
to game programming.

I also am really cheese off with an error that means I have to
restart my lisp session after every run of a demo. I think it is
that I am not cleaning up opengl resources adfter I am done with
them and thus the next tiem I start creating opengl resources it
simply crashes. I'm going to go on a hunt for this as this bug is
utterly unacceptable.

Other that example 4 I have just made a couple of bug fixes in a
few files.
Cheers all.