Added Generics and Fixed Bugs -----------------------------
I have created the package vectors which provides
generic version of all the vector functions. It
should NOT be ':USE'd in any other package as it
shadows + - * / length and others from common lisp.
The package has a nickname of 'v' so it is nice to
use without having to ':USE' it.

For example:

(v:length (v:+ #v(1 2 3) #v(-4.8 3.22 0)))

will return the length of the vector you get when you
add the two vectors together.

Note that the reader macro #v allows you to make new
vectors in a really clean way and not have to worry
about the type of the number being passed.

If you need a generic way of making vectors which can
be funcall'd and apply'd then you can use v:make-vector.

In making the vectors package I found a few bugs in the
vector packages which are now fixed.

Also it seems that I have not lost the history of the
vector and matrix packages, I just need to use the
correct flag in 'git log' to see them.

I am working on a new example which will show the new
vector commands but I'd like to try some other things
before pushing that live.