This commit is a little smaller than usual. I've been tearing apart sdl's 'with-events' macro to see what they include in the main loop. I'm trying to make as thin a layer between the user and the code as possible do I feel that the 'with-events' macro has a little too much magic.
In 7.lisp I have ripped out the parts I need to make this
function in the same way as 6.lisp.

I am currently experimenting with time in the protocode
folder, and as soon as I have nailed that down I will
and player controls to 7.lisp (or prehaps another) example.