We have temporal lambdas! I've been having good fun thinking about time recently. Spurred on by needing time functions for the input system, I have been trying to imagine what the primitive components of time handling could be.
For some reason the phrase 'lexical time' closely followed by
'temporal lambda' popped into my head and so I have created the later
with the hope that it could be useful. To be honest this is all just
exporation and experimentation at the moment but I have a gut feeling
that this could be useful.

Stay tuned to find out what happens next!