Ok so I've reworked the temporal functions and macros. I have changed the temporal predicates to be the thign that emit the 'temporally-expired' condition. They now take a number rather than a time-source and return the time input if the condition has been met. There is no need to emmit t or nil state whether the predicate has expired as this is done by the condition. These changes mean that the temporal predicates can be chained.
The make-tlambda function has been removed in favour for a more
complex tlambda macro which now has a variable !time bound to the
current time in the time source. This allows you to write the
temporal predicate using it and also make use of the time within
the body of the tlambda.

For example
(tlambda (make-time-cache) (withinp 1000 5000 !time)
   (print !time))

Will print the current time in the time cache as while the tlambda
has existed for between 1 and 5 seconds.

I've also updated walk-replace in cepl-utils to allow the user to
provide a custom function to use to test the atom for replacing.

That all for now, sorry this note is a bit harder to read that usual, my brain is on a wonk