First commit of a new branch I'm making to change how buffers are represented withing cepl-gl.
Up until now I have been using just the buffer-id integer to
identify the buffer. One problem with this has been that I have
had to manage the layout of that data seperately. I have now
switched from a buffer just being the integer id to being a
struct which stores both the id and also the format.

I've gone through and tweaked all the functions I can see which use this and am now off to test it by rewriting all the examples
so that they make use of this new feature. The result should bemuch shorter, lispier code!

Wish me luck!