This still isn't tested yet and I know there are beefy bugs in here but i have made some significant progress.
I have glstructs working well weven though they have seriously
ugly code. I'll be honest that alot of this will need going
over once I'm nearly done!

One things that has been wonderful is being able to read data
back from the GPU. GPU-arrays are abstractiosn that provide
an array like interface to GPU memory. It is lovely to be able
to maps a GPU-array to a GL-array (a local one) and the access it
just like any other GL-array (using gl-aref and struct getters).

This will be HUGE for repl based iteritive development as you
don't have to keep track of the source data to be able to go and
make in-line changes to data on the GPU.

I am now working through the example code, making sure the
original opengl style codee still works and then addign examples
using this new system of data management.

Hopefully I'll be in a position to merge back into master in
a few days.

Wish me luck!