This is a hefty update (as it should be after such a lull)
base-maths has two more functions for comparing a float to 0.
They have ugly names and chances are I will want to relook at
these soon, but I needed them for the quaternions.

cepl-camera is still a stub of a file but I updated the
make-cam-clip-matrix function as it was ugly.

cepl-gl has had some minor tweaks to fix a spelling error and
an error in the generic specifier of the make-gpu-array function.
I'm using a 'eql' specifier which isa bit pants and I'm fairly sure
I can fix this so I will be revisiting this.
Also I reworked the naming and purpose of the make-shaders and
make-shader functions. Originally make-shaders took file names and
loaded and the shaders, but this a a function doing two things
which should be split. load-shader and load-shaders now load the
shader files and pass the contents to make-shader. maker-shader
compiles the shader source and returns opengl shader objects.
make-program now takes any number of shaders using the &rest
parameter rather than taking a list of shaders. (this has broken
all the other examples so I need to fix these in the next commit!
sorry about this for anyone following, it'll be up in about half
an hour)

cepl.asd has has matrices.lisp removed as it no longer exists

quaternions.lisp is the code for handling quaternions. This is
unfinished but I have some progress so I best commit early and
often.

I have added examples showing how to make use of cepl-gl's new
datatypes to make working with opengl lispy and clean.
of special note would be 0-containing-shaders.lisp in which a
colour triangle is shown on the screen in only 27 lines of lisp,
in fact the file also contains the source for both the vert and
frag shaders and all together is only comes to 53 lines of code.
The latest example is 8.lisp which shows the start of a lighting
example.

base-matrices has had it's m! macro cleaned up.

the vector and matrix creating reader macros have been removed
as the v! and m! macros feel more lispy and provide exactly the
same benefits. This means rather than:
#v(1 2 3) we can use (v! 1 2 3) and instead of:

#m(1 2 3                 (m! 1 2 3
   4 5 6     we can use      4 5 6
   7 8 9)                    7 8 9)

Right, thats all for now.