*Not Currently Working* I'm in process of writing the lisp->glsl translator and its going pretty well. I am working through a copy of the 330 spec and slowly adding all the functionality.
I needed to push as this was getting to be too large a changeset for me to be happy just having it on my computer.

Changes:
* Fixed a bug in cepl-gl where a generic function was refering
  to a class created later in the file.

* Added intesperse to the cepl-utils folder

* Added cglsl-core cglsl-maths cglsl-matrices cglsl-test and
  cglsl-vectors. These are going to hold functionality which will
  mirror that of the cepl meaning the user can happily copy code
  between gpu shaders and local functions and know they will work.

* Modified cglsl to keep this file for meta functionality of the
  cglsl stuff.

* Added generic matrices functionality. This did require tweaking
  matrix3 and matrix4 a little. I need to come back to those two
  and really spruce them up

* Tweaked a doco string

Right, like I said at the top this isn't a working push
unfortunatly, but I'm slowly chugging through the work now so it
shouldn't be too many weeks before I get my first shaders written
in lisp.
Take care all!