*WARNING NOT CURRENTLY WORKING* As you can see from the log I have been focusing on cepl-gl. The major thing is that I have found out how I want to handle foreign types: I have defined foreign bool, float, int and uint based on the cffi basic types. From these I have defined all the basic glsl types (except samplers for now). The reason for redefining the 4 most primitive types is that then all the glsl types are under the cepl-gl package and so the programmer doesnt have to think 'hmm was the float part of cffi or cepl-gl?' though obviously they can if the want to. Structs can also be used in other structs. For example the following:
(defglstruct vert
  (position vec3)
  (color vec4))

(def poly
  (vectices vert :count 3))

Is totally valid.
All structs produce getters and setters of their own as well as for
their slots which makes it possible to use a slightly experimental
style of setting their contents.
For example using the structs defined above:

(setf (poly-vert (aref-gl *example-poly 1)) '((v! 1 2 3)
                                              (v! 4 5 6 7)))

will set the first vert in the poly's vert array to the contents
of the list where the contents are assigned in order, slot-wise.

Next I am back working on the defshader and defprogram code, trying
to use these types to make setting streams and uniforms both as
efficient and as easy as possible. This will take a while
especially as I am moving soon, but it will be great when it is
done. At that milestone I will finish the lisp->glsl translator and
rewrite all the examples to show how well it can work.

Goodnight all!