Well, I had a great time away and now I'm ripping out the old attempt at generating glsl and replacing it with varjo. Varjo is my lisp->glsl translator and while it still is in need of a hefty refactor the interface is solid enough to be worth integrating with Cepl.
This is mostly a reworking of cepl-gl-types and cepl-gl to start down the road of getting this
to work. I have gl-arrays and gpu-arrays working well, I now need to look at generating vaos,
this turns out to be a bit of a pain in the backside! It needs to be able to handle any glsl
type and also any type created by the user using defglstruct, getting with right is taking time.

Okidokey, ciao for now