Switching from lispbuilder to lispbuilder-mini.
Lispbuilder-mini is still in very early stages but as no-one
really is playign with this code yet I can afford to dog-food
this in trunk for a while.

Lispbuilder-mini will give us a much more trimmed down interface
with only the window and gl-context to worry about. I am
actively going through all of lispbuilder-sdl and removing
anything that is not related to pure opengl game coding.
2D in modern games is using shaders and such so all surfaces
and their commands are ignored unless it's related to the
display surface. I will be expanding the commands around the
gl context eventually as well as I want better control of it.

Follow lispbuilder-mini for details on what is going on.