gpu-streams have been renamed vertex-streams...because that is what they are. gpu-stream implies some kind of generic datatype when the only data we 'stream' into a shader is a vertex-array.
Also I have changed make-vertex-stream-from-gpu-arrays to just be make-vertex-stream. 90% of
the time this is how they will be created anyway so this is the default with
make-raw-vertex-stream be the way to make just the struct.

This has been bugging me for a week now so I'm glad I've got it cleaned up.