Adding varjo as a submodule
This brings varjo as a part of the cepl source. The idea here is to
reduce the steps needed to start using cepl.It just makes sense to have
varjo and cepl together in the same project.

However to preserve varjo's status as a independent project we are
bringing it as a git submodule. Once the repository is cloned one has to
simply do a

$ git submodule update --init