Part way through rewriting defpipeline
Need to do def-pipeline-init next, this one will be the tricky bit as
it is doing the actual compiling. We need to make lots of helpers to
make this thing smaller.

defpipeline is totally macro based now which means expand it will be
more incremental and pleasant.

Most of the work for gpu-function specs is in place so we just
essentially we just  need to compose the code and throw it at varjo. We
shall see how easy that ends up being :)