Clean up vector & matrix packages
The vector and matrix package are not meant to be 'use'd. So it made
no sense that we had v* v/ v+ v- etc. Now we shadow the symbols from
cl and named more sensibly.

Also finally killed that shitty apply-across-elements macro, it was
my first ever macro in lisp and it sucked :D

Removed all the stupid generic stuff from matrices.lisp