Make dependent function finding code more robust
While not ideal it does work so Ι have spent a little time making it
work better.

The logic boils down to:
"We need to know the function dependencies before we compile".

This sucks a bit but we can fully macroexpand, walk and handle the
special functions in a cond. This turns out to not be too bad as many of
the special funcs are essentially function calls.