Kill the code walker
The code walker was not robust. Have removed it in favor of using macros
to inject functions. This relies on varjo's deduplication of functions.

Right now it easts missing function errors so we should look at a
gpu-func-spec's 'missing-dependencies slot to see if it is incomplete.
We can then throw an exception if any incomplete func is used in a pipeline