can now detect gpu space scopes!
Been making progress on gpu spaces. Most of the work has been in
varjo. But things are looking good here too.

I can now perform a varjo pass and then analyze the ast for places
where the position being returned (from an expression) is from a
different space than the current one.

This is great as these are the places where we need to upload transform
matrices.