Can now compile basic shaders using spaces :D
I need to remove unused uniforms as they are causing issues.

It boils down to the fact that uniforms are only removed in the
transform pass, and the transform pass is only run on nodes that match
the filter. So if a uniform isnt used then there is nothing to match the
filter and thus nothing to remove the uniform.

This should be easy enough though, we can just check if each uniform's
flow-id is returned by any ast-node.