add internal pass to remove unused uniforms
yay! now we can throw all the uniforms we like at this and remove the
stuff we dont need. AWESOME!

I wonder how hard it will be to make a dead code remover. I'll ponder
that for a bit.. at it's heart it will be quite similar in that it will
be querying for flow-ids that are never used again, but it will need
some extra logic