I understand how spaces will work now :D
Finally got the last piece of the puzzle in my head so now I see how
this feature will look and the language used to talk about it.

I now need to implement a dumb version so I can test stuff. After that
we will need something that actually performs well, but holy crap I'm so
relieved!

`protocode/space/the-breakdown.txt` is the place to see the overall
picture. I will be filling out details there.