Fix g-func/pipeline compile order: WOOOO! :D
What a bloody mission that was. Took ages to trace the issue and then be
satisfied with the conclusion. So the two main changes are:

- upload gfunc specs immediately (at macro expansion time)
- dont use the 'change signature' to avoid recompiles..just do it. We
  can make this better later.

Also I added a verbose option to the %touch-* function. This will return
some details include what the uniform assigner & uniform transforms were
at the point the pipeline was compiled. This allows me to look at what
is inside the init and dispatch functions without macroexpanding (which
is important as expanding has side effects :p ahh I do abuse this
language)