Narrowing in on new space api
So the space routes system is too memory hungry but it will serve it's
purpose for now.

I pretty much have the new api now for making spaces.

Any one way spatial relationship doesnt need to be in the route system
as it has only 1 option to get anywhere. We call these the model-spaces
as that is what they will be used for (may change that name). You can
make as many of these as you like without incurring the n² memory issue.

We have 3 kinds of spaces: model, hierarchical, and relational

- hierarchical & model space can be a root for child spaces.
- only relational spaces can be a child space
- hierarchical spaces the only spaces in the route graph
- get-transform works between all space types (as long as the route
  makes sense)
- as model space is 1 directional it can be transformed to... I could
  change this though by finding the route to the 1 neighbour hmmm

I still need to add a way to upgrade a model space to a relational
space (which happens if another space adds you). And i also need to let
you update your relationship transforms.

In other news I also learned that case works by identity (#'eq) so using
it for any numeric was DUUUUMB. Fixed that now.