Add with-rendering-via
so here's a kicker, we want to get things to clip space via some
cam->clip space. But which one? When map-g does transforms it uses
get-transform which will take the shortest path (which is always the
identity-cam space) so we need a way to control this.

For now that way is with-rendering-via which checks to see if you are
going through clip-space and then enforces a path through a particular
space. madness