fix some cross pass bugs
The compile had issues with adding the cross space transform as in this
case.

```
     (defun-g first-vert ((vert vertex) &uniform (mspace space-g))
       (in *clip-space*
	 (progn
	   (progn
	     (in mspace (p! (pos vert) 1.0))))))
```
It thought every `progn` was a cross-space node. So we ended up with a
model->clip multiply at every one. The problem is that changing the
correct node doesnt tell the rest of the ast nodes that things have
changed. To fix this we do a depth first search for the problem nodes
and after changing each on we recompile. This way all the flow data
is correct and we can keep working.

I'm pretty sure there are still bugs in this but it fixes the main
graphical issue I was having so yay.

This now means that model->relational and relational->relational routing
is working correct and we get good results. (it was a total SHIT to
track all this down :D)