more docs, worked out viewport dimensions issue
Was having issues for ages with the naming and types of the dimensions
of the viewport. Having them as a list made it really clean for passing
to arrays, but having as a vector was so clean for other tasks.

The solution is as follows:
- Dimensions always returns a list
- Resolution always returns a vecn

Much nicer, now I need to make this consistent throughout CEPL