Argg I just cant stop tweaking! (buffers)
Buffers should be a collection of buffer-backed gpu-arrays.
This gives the symmetry with textures, better performance and a more
consistant story.

I have started tearing up the old code to make the new buffers work.
Going slowly but well.

Will need some kind of subseq-g that allows changing the type as the
element-type of the base arrays is always :uint8
Also how do we handle ownership/freeing etc