integrate external-funcs
This turned out to be tricky as the name of the function is no longer
enough, we also need to know the in-arg's types. I added a func-key type
that we can use to identify functions and have been updating code to use
it.

Also def-g-> had to be expanded to allow more styles of stage definition
as we also need to be able to define types here too.