make-texture now can take a pixel-format
This change comes from a better understanding of pixel-formats.

I realized had put too much import on image-format. The three things we
need to care about are:

- element-type: the type of the foreign data
- pixel-format: the layout of the components in that foreign data. This
                is important whe we have bgr textures etc
- image-format: the layout on the gpu. OpenGL converts to this for us.

Before I had been making sure the foreign data type was compatible with
(image-format->lisp-type img-format). However what we actually needed
was to check if the element-type and pixel-format we compatible.

This fixes this and also lets the user provide a pixel-format to cover
cases where CEPL would pick a more standard pixel-format (instead of a
packed one for example).