Work around 'multi defun in file' bug
So we generated defuns for our gpu-funcs, however we allow overloading
now so we were getting multiple defuns with the same name in the same
file (which is an error according to the spec).

They also can have non-congruent lambda lists so we couldnt just emit
defmethod's instead. Glsl requires this, for proof of this see [0]

For now we get around this by just using:
```
(setf (symbol-function ,n) ...)
```
This gets us out of trouble and keeps things working enough for now
whilst I work on a better solution.

[0]
```
(let ((g (groupby:groupby #'second glsl-spec:*functions* :test #'equal)))
  (loop for (n fs) in g collect
       (cons n (remove-duplicates (mapcar (lambda (x) (length (sixth x))) fs)))))
```