cache gl version info on making context
This avoids issues when compiling a gfunc from a thread other than the
GL thread. The GL context is not needed in this initial lisp->glsl
compile but we do need the glsl version.

To that end we queried the gl-context, but the context could potentially
be on another thread. By touching #'major/minor-version we ensure the
version is cached.